App: fix recompute optimization
This is a regression caused by 603502b47e088c4a18.

The key is to not set DocumentObject::_recompute which will force
recomputing the object. Instead, set status flag Touched and Enforced,
which enables recompute optimization in Document::_recomputeFeature() to
only recompute if there is actual change in any property.

Fixes realthunder/FreeCAD_Assembly3#1119