App/Gui: change child claiming behavior in GroupExtension
Add hidden property 'ClaimAllChildren' to control child claiming
behavior in GroupExtension. If this property is false, then the
GroupExtension behaves the same as the old GeoFeatureGroupExtension,
which only claim children objects in tree view that are not claimed by
other children. The code is reused from
(Viewprovider)GeoFeatureGroupExtension.

Default setting for new groups is exposed to
    'Preference -> General -> Objects'.

Fixes realthunder/FreeCAD_Assembly3#1139