Part: add parameter to control parallel run
With BaseApp/Preferences/Mod/Part/ParallelRunThreshold

With OCC version < 7.5, this parameter control whether to run
parallel with there is only two shapes doing boolean operation, and
total faces involved exceed this threshold.

With OCC version 7.5 onwards, always enable parallel run if this
parameter is greater than zero.

The original purpose of this parameter is to avoid parallel run for
small objects, because older OCC used to create and destroy threads to
run in parallel, which creates lots of overhead for large amount of
small operations. It has since been changed to use thread pools.