Gui: add new API Gui::isValidBBox()
libCoin3D SbBox3f::isEmpty() does not check for NaN in the coordinate
which may happen because of some code in FreeCAD, e.g.
Base::Vector3d::GetAngle() returns NaN when the input is a zero length
vector.

The new API is added to check for NaN to improve robustness in bound box
checking.