Part: add support for base attachment in AttachExtension
The base attachment specifies which geometry (mostly of the host
object's own shape) to act as a base for attachment.

The implementation relies on a set of dynamic properties (similar to the
existing ones in AttachExention) to be installed on demand when the user
enables base attachment, through AttachExention::getInitedProperties()
or through Python API AttachExentionPy.initBaseProperties(). The new
properties are

* BaseAttachment (counter part of the existing AttachmentSupport). This
  is a property link with hidden scope so that it can contain link to
  the host object which would normally cause a dependency loop
* BaseMapMode (v.s. MapMode)
* BaseMapReversed (v.s. MapReversed)
* BaseMapPathParameter (v.s. MapPathParameter)

An extra task panel is added for the base attachment.