Tools: fix Python extension class generator
Support multi-inheritenace twin class

static_cast PyObjectBase::_pcTwinPointer (which is declared as void *)
directly to twin class may cause error if the twin class or any of its
parent class has multiple inheritenace that is not the first
inheritenace.

The solution is to generate a common class member function
getTwinPointer() that will call the parent class's getTwinPointer() to
chain the pointer casting.